+++ examples/04-normalizr-github-paged-project-search/lib/search-page/schema.js
@@ -1,4 +1,5 @@
-import { Schema, arrayOf } from 'normalizr';
+import { schema } from 'normalizr';
 
 // we aren't worrying about a complex schema... just demonstrating the ability to normalize based on a schema
-export default arrayOf(new Schema('repository'));
+const RepositorySchema = new schema.Entity('repository');
+export default new schema.Array(RepositorySchema);