+++ examples/05-normalizr-github-paged-project-search-and-viewer/lib/project-details-page/schema.js
@@ -1,4 +1,7 @@
-import { Schema, arrayOf } from 'normalizr';
+import { schema } from 'normalizr';
 
 // we aren't worrying about a complex schema... just demonstrating the ability to normalize based on a schema
-export default arrayOf(new Schema('repository', { idAttribute: 'full_name' }));
+const RepositorySchema = new schema.Entity('repository', {
+  idAttribute: 'full_name'
+});
+export default new schema.Array(RepositorySchema);