+++ src/index.js
@@ -2,10 +2,14 @@ import Model from './model';
 import reducer from './model-reducer';
 import modelProvider from './model-provider';
 import reduxEffectsActionCreator from './action-creator/redux-effects';
+import fetchJSONMiddleware from './fetch-json-middleware';
+import dispatchPromise from './dispatch-promise';
 
 export {
   Model,
   reducer,
   modelProvider,
-  reduxEffectsActionCreator
+  reduxEffectsActionCreator,
+  fetchJSONMiddleware,
+  dispatchPromise
 };